improve maintainability
