+++ src/rules/unused_column_in_cte/visitors/mod.rs
@@ -1,5 +1,4 @@
 mod cte_visitor;
-mod distinct_visitor;
 mod pivot_visitor;
 mod qualify_visitor;
 mod select_star_visitor;
@@ -7,9 +6,8 @@ mod select_visitor;
 mod where_visitor;
 
 pub use cte_visitor::CteVisitor;
-pub use select_visitor::SelectVisitor;
-pub use where_visitor::WhereVisitor;
-// DistinctVisitor not exported - DISTINCT doesn't make all CTE columns used
 pub use pivot_visitor::PivotVisitor;
 pub use qualify_visitor::QualifyVisitor;
 pub use select_star_visitor::SelectStarVisitor;
+pub use select_visitor::SelectVisitor;
+pub use where_visitor::WhereVisitor;
